+++ packages/werkbank/lib/src/addons/src/knobs/src/knobs/int_knob.dart
@@ -32,6 +32,29 @@ extension IntKnobExtension on KnobsComposer {
       valueFormatter: valueFormatter,
     );
   }
+
+  /// Creates an integer knob controlled by a text field in the UI.
+  ///
+  /// {@macro werkbank.knobs.label}
+  ///
+  /// {@macro werkbank.knobs.regularInitial}
+  ///
+  /// {@template werkbank.knobs.intField}
+  /// The [min] and [max] parameters define the range of allowed values.
+  /// {@endtemplate}
+  WritableKnob<int> intField(
+    String label, {
+    required int initialValue,
+    int? min,
+    int? max,
+  }) {
+    return customField<int>(
+      label,
+      initialValue: initialValue,
+      parser: (input) => _intInputParser(input, min: min, max: max),
+      formatter: _defaultFormatter,
+    );
+  }
 }
 
 /// {@category Knobs}
@@ -63,6 +86,54 @@ extension NullableIntKnobExtension on NullableKnobsComposer {
       valueFormatter: valueFormatter,
     );
   }
+
+  /// Creates a nullable integer knob controlled by a text field in the UI.
+  ///
+  /// {@macro werkbank.knobs.label}
+  ///
+  /// {@macro werkbank.knobs.nullableInitial}
+  ///
+  /// {@macro werkbank.knobs.intField}
+  WritableKnob<int?> intField(
+    String label, {
+    required int initialValue,
+    bool initiallyNull = false,
+    int? min,
+    int? max,
+  }) {
+    return customField<int>(
+      label,
+      initialValue: initialValue,
+      parser: (input) => _intInputParser(input, min: min, max: max),
+      formatter: _defaultFormatter,
+      initiallyNull: initiallyNull,
+    );
+  }
 }
 
 String _defaultFormatter(int value) => value.toString();
+
+InputParseResult<int> _intInputParser(
+  String input, {
+  int? min,
+  int? max,
+}) {
+  final trimmedInput = input.trim();
+  if (trimmedInput.isEmpty) {
+    return InputParseError('Input required');
+  }
+
+  final value = int.tryParse(trimmedInput);
+  if (value == null) {
+    return InputParseError('Invalid integer format');
+  }
+
+  if (min != null && value < min) {
+    return InputParseError('Must be ≥ $min');
+  }
+  if (max != null && value > max) {
+    return InputParseError('Must be ≤ $max');
+  }
+
+  return InputParseSuccess(value);
+}