+++ packages/werkbank/lib/src/addons/src/knobs/src/knob_types/nullable_knob.dart
@@ -9,6 +9,9 @@ typedef NullableKnobBuilder<T extends Object> =
       ValueNotifier<T> valueNotifier,
     );
 
+@Deprecated('Use NullableKnobsComposer instead')
+typedef NullableKnobs = NullableKnobsComposer;
+
 extension type NullableKnobsComposer(KnobsComposer _knobs) {
   WritableKnob<T?> makeNullableKnob<T extends Object>(
     String label, {