+++ packages/werkbank/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Added a `Werkbank System` theme, which will always match your system. It's also default now.
+
 # 0.16.0
 - Add `c.background.colorBuilder(...)` and `c.background.widgetBuilder(...)` to allow easier use of theme colors in backgrounds.
 - Fix that pointer could interact with use case while trying to change constraints.