+++ packages/werkbank/CHANGELOG.md
@@ -8,6 +8,7 @@
   - `c.knobs.customSlider<T>(...)` (for custom slider controlled knobs)
   - `c.knobs.customSwitch<T>(...)` (for custom switch controlled knobs)
   - And the corresponding nullable variants.
+- Add `intInput` and `doubleInput` knobs.
 - Add `falseLabel` and `trueLabel` parameters to `c.knobs.boolean(...)` to allow customizing the labels of the switch.
 - Rename `NullableKnobs` to `NullableKnobsComposer`. **(BREAKING CHANGE)**